Replace `Get<float>` by `Get<double>` in property parsing
With latest version of Boost (1.88), `Get<float>` raises "RuntimeError: bad
lexical cast: source type value could not be interpreted as target"
exception for floating point values that don't exactly fit `float` precision.

Such values can be very innocent-looking: for instance, 39.5978 cannot be
stored as is in a float, but will be stored as 39.597801, what will make
Boost raise, and give the impression the problem is random.

To fix it, the only solution was to parse floating point values as
double and then narrow it to float.